chore(test): move all tests under `tests/`
