+++ src/lib.rs
@@ -110,37 +110,6 @@ pub trait Node {
     }
 }
 
-#[rustfmt::skip]
-impl Node for BitcoinD {
-    fn get_name() -> &'static str { "bitcoind_v_31_0" }
-
-    fn get_chain_tip(&self) -> Result<u32, Error> { self.get_chain_tip() }
-
-    fn get_block_hash(&self, height: u32) -> Result<BlockHash, Error> { self.get_block_hash(height) }
-}
-
-#[rustfmt::skip]
-impl Node for UtreexoD {
-    fn get_name() -> &'static str { "utreexod_v_0_5_0" }
-
-    fn get_chain_tip(&self) -> Result<u32, Error> {
-        let height = self.get_chain_tip()?;
-        if height == 0 {
-            return Err(
-                Error::UnexpectedResponse("utreexod is at genesis, proof index not yet available".to_string())
-            );
-        }
-        self.get_block_uproof(height)?;
-        Ok(height)
-    }
-
-    fn get_block_hash(&self, height: u32) -> Result<BlockHash, Error> { self.get_block_hash(height) }
-
-    fn poll_interval() -> Duration { 2 * POLL_INTERVAL }
-
-    fn wait_timeout() -> Duration { 2 * WAIT_TIMEOUT }
-}
-
 /// Poll a [`Node`] until its chain height reaches `height`, then return.
 ///
 /// Panics if the node does not reach `height` within [`Node::wait_timeout`].
