+++ src/utreexod/mod.rs
@@ -500,88 +500,11 @@ pub fn get_utreexod_path() -> Result<PathBuf, Error> {
     }
 }
 
-#[cfg(all(test, feature = "utreexod_0_5_0"))]
-mod test {
-    use crate::wait_for_height;
-
-    use super::*;
-
-    /// Verify that [`UtreexoD`] starts successfully and
-    /// exposes its PID, working directory, and P2P socket.
-    #[test]
-    fn test_utreexod_starts() {
-        let bin_path = get_utreexod_path().unwrap();
-        let utreexod = UtreexoD::from_bin(bin_path).unwrap();
-
-        println!("PID: {}", utreexod.get_pid());
-        println!("Working Directory: {:?}", utreexod.get_working_directory());
-        println!("P2P Socket: {}", utreexod.get_p2p_socket());
-    }
-
-    /// Verify that `generate` mines the requested number of blocks.
-    #[test]
-    fn test_utreexod_generate() {
-        let utreexod = UtreexoD::new().unwrap();
-
-        let height = utreexod.get_chain_tip().unwrap();
-        assert_eq!(height, 0);
-
-        utreexod.generate(10).unwrap();
-
-        let height = utreexod.get_chain_tip().unwrap();
-        assert_eq!(height, 10);
-    }
-
-    /// Verify that [`UtreexoD::get_block_hash`] returns the correct hash for a given height.
-    #[test]
-    fn test_utreexod_get_block_hash() {
-        let utreexod = UtreexoD::new().unwrap();
-
-        let block_hashes = utreexod.generate(10).unwrap();
-
-        let last_block_hash = utreexod.get_block_hash(10).unwrap();
-
-        assert_eq!(last_block_hash, *block_hashes.last().unwrap());
-    }
-
-    /// Verify that two nodes can connect to each other via `add_peer`,
-    /// and that the peer count reflects the new connection on both sides.
-    #[test]
-    fn test_utreexod_addnode() {
-        let utreexod_alpha = UtreexoD::new().unwrap();
-        let utreexod_beta = UtreexoD::new().unwrap();
-
-        assert_eq!(utreexod_alpha.get_peer_count().unwrap(), 0);
-        assert_eq!(utreexod_beta.get_peer_count().unwrap(), 0);
-
-        utreexod_beta
-            .add_peer(utreexod_alpha.get_p2p_socket())
-            .unwrap();
-
-        assert_eq!(utreexod_alpha.get_peer_count().unwrap(), 1);
-        assert_eq!(utreexod_beta.get_peer_count().unwrap(), 1);
-    }
-
-    /// Verify that blocks mined on one node propagate to a peer.
-    #[test]
-    fn test_utreexod_blocks_propagate() {
-        let utreexod_alpha = UtreexoD::new().unwrap();
-        let utreexod_beta = UtreexoD::new().unwrap();
-
-        utreexod_alpha.generate(21).unwrap();
-
-        assert_eq!(utreexod_alpha.get_chain_tip().unwrap(), 21);
-        assert_eq!(utreexod_beta.get_chain_tip().unwrap(), 0);
-
-        utreexod_alpha
-            .add_peer(utreexod_beta.get_p2p_socket())
-            .unwrap();
-
-        wait_for_height(&utreexod_beta, 21).unwrap();
-        assert_eq!(utreexod_beta.get_chain_tip().unwrap(), 21);
-
-        utreexod_beta.generate(21).unwrap();
-        wait_for_height(&utreexod_alpha, 42).unwrap();
-        assert_eq!(utreexod_alpha.get_chain_tip().unwrap(), 42);
+impl Drop for UtreexoD {
+    /// Kills the `utreexod` process.
+    ///
+    /// Errors from `kill` are silently discarded so that `Drop` never panics.
+    fn drop(&mut self) {
+        let _ = self.process.kill();
     }
 }
